GameTimer added And CardGameTester
Added a file to test the war game
ultimately a non player is always made
this is because the cards are being added to the deck as a win, instead of a score being added up.
card deck != score

Updated C# file table to count down from a specificed time.

~ TimeLeft ~The interval of count down can be changed (60 = 1miunte, 1 = 1 second)
~ GameTime~ The number of minutes that the game is allowed to be played (10 = 10minutes, 1 = 1minute)